pyproject.toml
[project]
name = "standardragdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.5",
    "faiss-cpu>=1.13.0",
    "ipykernel>=7.1.0",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langchain-groq>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "pymupdf>=1.26.6",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
